Rename generator to httpHandler
