+++ cmd/generator/generator.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/superhacker2002/parentheses/internal/controller/generator"
+	"github.com/superhacker2002/parentheses/internal/controller/httphandler"
 	"github.com/superhacker2002/parentheses/internal/parentheses"
 	"log"
 	"math/rand"
@@ -10,7 +10,7 @@ import (
 )
 
 func main() {
-	handler := generator.New(parentheses.New())
+	handler := httphandler.New(parentheses.New())
 	handler.SetRoutes()
 	rand.Seed(time.Now().UnixNano())
 	log.Fatal(http.ListenAndServe("localhost:8080", nil))
