+++ internal/controller/httphandler/httphandler_test.go
@@ -1,4 +1,4 @@
-package generator
+package httphandler
 
 import (
 	"github.com/stretchr/testify/assert"
